requirements.txt
allure-behave
allure-pytest
Appium-Python-Client
behave
config
iniconfig
openpyxl
pandas
parse
pytest
python-dotenv
requests
sauceclient
selenium
webdriver-manager
beautifulsoup4
Image
Pillow
python-docx
docx
pyautogui
xlrd
pyodbc
